[Refactor] Move the SettingsComponent to Collections
