+++ app/components/collections/settings_component.html.erb
